+++ src/REvoDesign/editor/monaco/server.py
@@ -15,7 +15,7 @@ from fastapi.staticfiles import StaticFiles
 from pydantic import BaseModel
 
 from REvoDesign import ConfigBus
-from REvoDesign.basic.abc_singleton import SingletonAbstract
+from REvoDesign.basic.server_monitor import ServerControlAbstract
 from REvoDesign.tools.ssl_certificates import SSLCertificateManager
 
 from ...logger import ROOT_LOGGER
@@ -276,7 +276,7 @@ async def save_file(
         return JSONResponse(content={"error": f"Failed to save file: {str(e)}"}, status_code=500)
 
 
-class ServerControl(SingletonAbstract):
+class ServerControl(ServerControlAbstract):
     """
     A singleton class that manages the Monaco backend server lifecycle.
 
@@ -297,9 +297,9 @@ class ServerControl(SingletonAbstract):
     """
 
     def singleton_init(self):
-        self.server_thread = None  # WorkerThread instance
+        self.server_thread: WorkerThread = None  # type: ignore # WorkerThread instance
         self.is_running = False
-        self.server = None  # Uvicorn Server instance
+        self.server: uvicorn.Server = None  # type: ignore # Uvicorn Server instance
         self.config_store = ConfigStore()
 
     def start_server(self):
@@ -373,23 +373,3 @@ class ServerControl(SingletonAbstract):
             self.server_thread.interrupt()
         self.is_running = False
         distruct_token()
-
-    def _run_server(self):
-        """
-        The function executed in the worker thread.
-        """
-        if self.server:
-            self.server.run()
-
-    def _on_server_result(self, result):
-        """
-        Handle results from the WorkerThread.
-        """
-        print(f"Server result: {result}")
-
-    def _on_server_finished(self):
-        """
-        Handle the completion of the WorkerThread.
-        """
-        self.is_running = False
-        print("Server thread finished.")