+++ src/REvoDesign/shortcuts/tools/mutation_effect_predictors.py
@@ -16,15 +16,23 @@ from REvoDesign.common.mutant import Mutant
 from REvoDesign.common.mutant_tree import MutantTree
 from REvoDesign.tools.mutant_tools import (extract_mutants_from_mutant_id,
                                            quick_mutagenesis)
-from REvoDesign.tools.utils import timing, get_cited, require_installed
-
+from REvoDesign.tools.utils import get_cited, require_installed, timing
 
 logging = ROOT_LOGGER.getChild(__name__)
 
 RUN_MODE_T = Literal["single", "additive", "epistatic"]
 
+
 @require_installed
 class ThermoMpnnPredictor(ThirdPartyModuleAbstract):
+    """
+    ThermoMpnnPredictor class for predicting the thermodynamic stability effects of protein mutations.
+    It utilizes the ThermoMPNN model to analyze protein structure and sequence to predict stability changes due to mutations.
+
+    Attributes:
+        name (str): Name of the predictor, set to 'ThermoMPNN'.
+        installed (bool): Indicates if the 'thermompnn' package is installed.
+    """
     name: str = 'ThermoMPNN'
     installed: bool = is_package_installed('thermompnn')
 
@@ -39,23 +47,49 @@ class ThermoMpnnPredictor(ThirdPartyModuleAbstract):
                  distance: float = 5.0,
                  ss_penalty: bool = False,
                  device: str = 'cpu'):
+        """
+        Initializes the ThermoMpnnPredictor with the given parameters.
 
+        Parameters:
+            pdb (str): Path to the PDB file of the protein.
+            save_dir (Optional[str]): Directory to save the output files. Defaults to None.
+            prefix (str): Prefix for the output files. Defaults to 'thermompnn_ssm'.
+            chains (Optional[List[str]]): List of chains to consider. Defaults to None.
+            mode (RUN_MODE_T): Mode of operation, either 'single' or 'batch'. Defaults to 'single'.
+            batch_size (int): Batch size for processing. Defaults to 256.
+            threshold (float): Threshold for some processing criteria. Defaults to -0.5.
+            distance (float): Distance threshold for some calculations. Defaults to 5.0.
+            ss_penalty (bool): Whether to apply secondary structure penalty. Defaults to False.
+            device (str): Device to run the model on, either 'cpu' or 'gpu'. Defaults to 'cpu'.
+        """
         self.prefix = prefix
         self.mode = mode
 
         from thermompnn import ThermoMPNN
         if save_dir and prefix:
+            # Construct the save prefix and create the directory if it doesn't exist
             self.save_prefix = os.path.join(save_dir, prefix)
             os.makedirs(os.path.dirname(self.save_prefix), exist_ok=True)
         else:
             self.save_prefix = ''
 
+        # Load the protein sequence from the PDB file
         self.sequence = RosettaPyProteinSequence.from_pdb(pdb)
+        # Initialize the ThermoMPNN application with the provided parameters
         self.app = ThermoMPNN(pdb, self.save_prefix, chains, mode, batch_size, threshold, distance, ss_penalty, device)
 
     @get_cited
     def run(self) -> pd.DataFrame:
+        """
+        Runs the ThermoMPNN prediction and returns the results as a DataFrame.
+
+        Returns:
+            pd.DataFrame: DataFrame containing the prediction results with columns 'ddG' and 'Mutation',
+                          or 'ddG', 'Mutation', and 'dist' depending on the mode.
+        """
+        # Process the protein and get the results as a DataFrame
         df = self.app.process(save_csv=bool(self.save_prefix))
+        # Rename the DataFrame columns based on the mode
         if self.mode == 'single':
             df.columns = ['ddG', 'Mutation']
         else:
@@ -64,6 +98,16 @@ class ThermoMpnnPredictor(ThirdPartyModuleAbstract):
 
     @staticmethod
     def mutant_name2mutant(mutant_id: str, sequences: RosettaPyProteinSequence) -> Mutant:
+        """
+        Converts a mutant ID string into a Mutant object.
+
+        Parameters:
+            mutant_id (str): String representation of the mutant.
+            sequences (RosettaPyProteinSequence): Protein sequence object.
+
+        Returns:
+            Mutant: Mutant object created from the mutant ID.
+        """
         return extract_mutants_from_mutant_id(
             mutant_string=mutant_id,
             sequences=sequences,
@@ -71,7 +115,18 @@ class ThermoMpnnPredictor(ThirdPartyModuleAbstract):
         )
 
     def df2mutant_tree(self, df: pd.DataFrame, sorted_by: Literal['prefix', 'positions'] = 'prefix') -> MutantTree:
+        """
+        Converts a DataFrame of mutation predictions into a MutantTree object.
+
+        Parameters:
+            df (pd.DataFrame): DataFrame containing mutation predictions.
+            sorted_by (Literal['prefix', 'positions']): Sorting criterion for the mutant tree. Defaults to 'prefix'.
+
+        Returns:
+            MutantTree: MutantTree object containing the mutation predictions.
+        """
         mutant_tree = MutantTree()
+        # Iterate over each row in the DataFrame to create Mutant objects and add them to the MutantTree
         for i, row in df.iterrows():
             score: float = row['ddG']
             mutation: str = row['Mutation']
@@ -80,6 +135,7 @@ class ThermoMpnnPredictor(ThirdPartyModuleAbstract):
             mutant.mutant_score = score
             mutant.wt_score = 0
 
+            # Add the mutant to the appropriate branch in the MutantTree
             mutant_tree.add_mutant_to_branch(self.prefix if sorted_by == 'prefix' else 'TMPNN_' +
                                              '_vs_'.join(str(m.position) for m in mutant.mutations), mutant.full_mutant_id, mutant)